requirements.txt
flask==2.0.1
kafka-python==2.0.2
cassandra-driver==3.25.0
Werkzeug==2.2.2
